feat: compact instruction format

+++ superr_vm/src/address.rs
@@ -1,1 +1,1 @@
-pub type MemoryAddress = usize;
+pub type MemoryAddress = u8;
